pathfinding, destruction of enemies/sprites etc, little bit of refactoring to fit the pathfinding in

+++ core/src/dev/adamhodgkinson/game/navigation/Arc.java
@@ -1,9 +1,9 @@
 package dev.adamhodgkinson.game.navigation;
 
 public class Arc {
-    final public byte weight;
+    final public short weight;
 
-    public Arc(byte weight) {
+    public Arc(short weight) {
         this.weight = weight;
     }
