+++ core/src/dev/adamhodgkinson/game/navigation/NavGraph.java
@@ -66,14 +66,14 @@ public class NavGraph {
 	/**
 	 * Directional from 1 to 2
 	 */
-	public Arc addLinearEdge(short x1, short y1, short x2, short y2, byte weight) {
+	public Arc addLinearEdge(short x1, short y1, short x2, short y2, short weight) {
 		final int index1 = coordToIndexMap[x1][y1];
 		final int index2 = coordToIndexMap[x2][y2];
 		adjacencyMatrix[index1][index2] = new Arc(weight);
 		return adjacencyMatrix[index1][index2];
 	}
 
-	public JumpArc addJumpEdge(short x1, short y1, short x2, short y2, byte weight, double xSpeed, double jumpSpeed) {
+	public JumpArc addJumpEdge(short x1, short y1, short x2, short y2, short weight, double xSpeed, double jumpSpeed) {
 		final int index1 = coordToIndexMap[x1][y1];
 		final int index2 = coordToIndexMap[x2][y2];
 		final JumpArc arc = new JumpArc(weight, xSpeed, jumpSpeed);
@@ -81,7 +81,7 @@ public class NavGraph {
 		return arc;
 	}
 
-	public void addBiDirEdge(short x1, short y1, short x2, short y2, byte weight, boolean isJump) {
+	public void addBiDirEdge(short x1, short y1, short x2, short y2, short weight, boolean isJump) {
 		addLinearEdge(x1, y1, x2, y2, weight);
 		addLinearEdge(x2, y2, x1, y1, weight);
 	}
